add order details page
